Enabling SSL
Now the server is either SSL or nothing. Certificates must be placed under

/etc/ssl/certs/server.crt
/etc/ssl/certs/server.key

otherwise the server will not be able to start. Pip requirements updated
with pyOpenSSL and related dependencies. In order for Pip to compile
some dependencies of pyOpenSSL, it is neccesary to install:

* libffi-dev
* libssl-dev
* openssl

The verification of the previous packages should be done using CMake, but
that is not the case right now.