Force compiler to use C++11.
Added comment in CMake regarding a bug in boost (at least <= 1.49) that
requires boost properties to be patched. Otherwise C++11 cannot be used.

All the code committed so far does not require C++11, but I will start
experimenting from now on, so either patching boost, or installing a new
version is necessary.