Added persistance to the MapStorage.
Now on start-up the map is loaded from disk, and whenever a new key/value
is added, the new version of the map is persisted. The location of the file
where the map is persisted is hardcoded to "/tmp/map". In the future the
location will be configuration based, and the name of the file will depend
on the customer executing the request.