Added Blowfish encryption of key and value using openssl
It is still in a very primitive state, but as a proof of concepts it
shows that it can be done. The enc/dec code is inside the MapStorage
which is not the  best option. I have to investingate how to move all
encryption related code to its own class, and think about the implications
of multiple threads using the same encryption/decryption class.